Diplom 3/requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
pytest==8.0.2
requests==2.31.0
selenium==4.18.1

